Create Due_Dil_Approvals_Explorer.JS
interacts with SC via infura. Uses the contract instance to call the proposals() function to get an array of proposals. It then loops through the proposals and uses the getVotesFor() and getVotesAgainst() functions to get the vote counts for each proposal. Then, it logs the proposals and their vote counts to the console.

Will add code to display the proposals and their votes in  the Web app (Due Dil portal)