Create rating.js
Sets up the Mongoose schema for the ratings, which includes fields for each of the due dil factors and their respective comments. When the form is submitted, the data is saved to the MongoDB database using the `save()` method.

Once the form is submitted, users will be redirected to a "Thanks for Submitting" page.

(install MongoDB and Node.js. Create new MongoDB database called "due-diligence". Update app.js connection string to match database config.)