Create ratingsController.js
This controller sets up a POST route for handling the form submission. When a user submits the form, the controller creates a new Rating instance with the submitted data, and then calls the save() method to save the rating to the database. If the save is successful, the controller redirects the user to a "Thanks for Submitting" page. If there is an error, the controller sends a 500 error response.